multi scene, fixes and cleanup

+++ wiesel/assets/shaders/geometry_shader.frag
@@ -36,7 +36,7 @@ struct LightPoint {
 layout (set = 0, binding = 0, std140) uniform Matrices {
     mat4 modelMatrix;
     mat3 normalMatrix;
-    float entityId;
+    uint entityId;
     // implicit 12 bytes padding to vec4 boundary
     vec4 colorTint;
     vec4 materialParams; // x=roughness, y=metallic, z=specular
@@ -76,7 +76,7 @@ layout(location = 6) in flat uint inFlags;
 layout(location = 7) in vec3 inViewDir;
 layout(location = 8) in vec3 inViewPos;
 layout(location = 9) in mat3 inTBN;
-layout(location = 12) in flat float inEntityId;
+layout (location = 12) in flat uint inEntityId;
 
 layout(location = 0) out vec4 outViewPos;
 layout(location = 1) out vec4 outWorldPos;
@@ -84,7 +84,7 @@ layout(location = 2) out float outDepth;
 layout(location = 3) out vec4 outNormal;
 layout(location = 4) out vec4 outAlbedo;
 layout(location = 5) out vec4 outMaterial;
-layout(location = 6) out float outEntityId;
+layout (location = 6) out uint outEntityId;
 
 
 vec3 getSurfaceNormal() {
