+++ wiesel/assets/shaders/shadow_shader.vert
@@ -7,7 +7,7 @@
 layout(set = 0, binding = 0, std140) uniform Matrices {
     mat4 modelMatrix;
     mat3 normalMatrix;
-    float entityId;
+    uint entityId;
     // implicit 12 bytes padding to vec4 boundary
     vec4 colorTint;
     vec4 materialParams;
